refactor(arch): Extract tables from applications utils
