+++ app/utils/applications.py
@@ -1,4 +1,3 @@
-import locale
 import logging
 import re
 from datetime import datetime
@@ -16,15 +15,7 @@ from sqlalchemy.sql.expression import true
 from app import models, parsers, serializers, util
 from app.i18n import get_translated_string
 from app.settings import app_settings
-from reportlab_mods import (
-    borrower_size_dict,
-    create_table,
-    document_type_dict,
-    sector_dict,
-    styleN,
-    styleSubTitle,
-    styleTitle,
-)
+from reportlab_mods import styleN
 
 logger = logging.getLogger(__name__)
 
@@ -56,25 +47,6 @@ class ERROR_CODES(Enum):
     DOCUMENT_VERIFICATION_MISSING = "DOCUMENT_VERIFICATION_MISSING"
     APPLICATION_LAPSED = "APPLICATION_LAPSED"
     APPLICATION_ALREADY_COPIED = "APPLICATION_ALREADY_COPIED"
-
-
-def format_currency(number, currency):
-    if isinstance(number, str):
-        try:
-            number = int(number)
-        except ValueError:
-            return "-"
-    locale.setlocale(locale.LC_ALL, "")
-    formatted_number = locale.format_string("%d", number, grouping=True)
-    return f"{currency}$ " + formatted_number
-
-
-def format_date(date_str):
-    if date_str == "None":
-        return "-"
-    date_object = datetime.strptime(date_str, "%Y-%m-%d %H:%M:%S")
-    formatted_date = date_object.strftime("%Y-%m-%d")
-    return formatted_date
 
 
 def update_data_field(application: models.Application, payload: parsers.UpdateDataField):
@@ -1334,88 +1306,6 @@ def check_if_application_was_already_copied(application: models.Application, ses
         )
 
 
-def create_borrower_table(borrower: models.Borrower, application: models.Application, lang: str):
-    """
-    Creates a table of borrower data.
-
-    :param borrower: The borrower's data.
-    :type borrower: Borrower
-
-    :param lang: The lang requested.
-    :type lang: str
-
-    :return: The generated table.
-    :rtype: Table
-    """
-
-    data = [
-        [
-            get_translated_string("MSME Data", lang),
-            get_translated_string("Data", lang),
-        ],
-        [
-            get_translated_string("Legal Name", lang),
-            Paragraph(borrower.legal_name, styleN),
-        ],
-        [
-            get_translated_string("Address", lang),
-            Paragraph(borrower.address, styleN),
-        ],
-        [
-            get_translated_string("National Tax ID", lang),
-            borrower.legal_identifier,
-        ],
-        [
-            get_translated_string("Registration Type", lang),
-            borrower.type,
-        ],
-        [
-            get_translated_string("Size", lang),
-            get_translated_string(borrower_size_dict[borrower.size], lang),
-        ],
-        [
-            get_translated_string("Sector", lang),
-            get_translated_string(sector_dict[borrower.sector], lang),
-        ],
-        [
-            get_translated_string("Business Email", lang),
-            application.primary_email,
-        ],
-    ]
-    return create_table(data)
-
-
-def get_pdf_file_name(application: models.Application, lang: str):
-    name = get_translated_string("Application Details", lang).replace(" ", "_")
-    filename = f"{name}-{application.borrower.legal_identifier}" + f"-{application.award.source_contract_id}.pdf"
-    return filename
-
-
-def create_pdf_title(title: str, lang: str, subtitle: bool = False):
-    """
-    Creates a table title for the application PDF.
-
-    :param title: The document's title.
-    :type title: str
-
-    :param lang: The lang requested.
-    :type lang: str
-
-    :param subtitle: If the title is subtitle.
-    :type subtitle: bool
-
-    :return: The generated title.
-    :rtype: Paragraph
-    """
-    if subtitle:
-        return Paragraph(
-            get_translated_string(title, lang),
-            styleSubTitle,
-        )
-    else:
-        return Paragraph(get_translated_string(title, lang), styleTitle)
-
-
 def create_table_cell(text: str, lang: str):
     """
     Creates a table cell for the application PDF.
@@ -1430,209 +1320,3 @@ def create_table_cell(text: str, lang: str):
     :rtype: Paragraph
     """
     return Paragraph(get_translated_string(text, lang), styleN)
-
-
-def create_application_table(application: models.Application, lang: str):
-    """
-    Creates a table of application information.
-
-    :param application: The application's data.
-    :type application: Application
-
-    :param lang: The lang requested.
-    :type lang: str
-
-    :return: The generated table.
-    :rtype: Table
-    """
-
-    data = [
-        [
-            get_translated_string("Financing Options", lang),
-            get_translated_string("Data", lang),
-        ],
-        [
-            get_translated_string("Lender", lang),
-            application.lender.name,
-        ],
-        [
-            get_translated_string("Amount requested", lang),
-            format_currency(application.amount_requested, application.currency),
-        ],
-    ]
-
-    if application.credit_product.type == models.CreditType.LOAN:
-        data.append(
-            [
-                get_translated_string("Type", lang),
-                get_translated_string("Loan", lang),
-            ],
-        )
-        data.append(
-            [
-                get_translated_string("Payment start date", lang),
-                format_date(str(application.payment_start_date)),
-            ],
-        )
-        data.append(
-            [
-                get_translated_string("Repayment terms", lang),
-                get_translated_string(
-                    "{repayment_years} year(s), {repayment_months} month(s)",
-                    lang,
-                    {
-                        "repayment_years": application.repayment_years,
-                        "repayment_months": application.repayment_months,
-                    },
-                ),
-            ],
-        )
-    else:
-        data.append(
-            [
-                get_translated_string("Type", lang),
-                get_translated_string("Credit Line", lang),
-            ],
-        )
-
-    if application.status == models.ApplicationStatus.COMPLETED:
-        data.append(
-            [
-                get_translated_string("Contract amount", lang),
-                format_currency(application.contract_amount_submitted, application.currency),
-            ]
-        )
-        data.append(
-            [
-                get_translated_string("Credit amount", lang),
-                format_currency(application.disbursed_final_amount, application.currency),
-            ]
-        )
-    return create_table(data)
-
-
-def create_documents_table(documents: List[models.BorrowerDocument], lang: str):
-    """
-    Creates a table of MSME information and documents.
-
-    :param documents: List of documents.
-    :type documents: List[Document]
-
-    :param lang: The lang requested.
-    :type lang: str
-
-    :return: The generated table.
-    :rtype: Table
-    """
-
-    data = [
-        [
-            get_translated_string("MSME Documents", lang),
-            get_translated_string("Data", lang),
-        ]
-    ]
-    for document in documents:
-        data.append(
-            [
-                get_translated_string(document_type_dict[document.type], lang),
-                document.name,
-            ]
-        )
-
-    return create_table(data)
-
-
-def create_award_table(award: models.Award, lang: str):
-    """
-    Creates a table of Open Contracting award data.
-
-    :param award: The award data.
-    :type award: models.Award
-
-    :param previous_awards: Previous award amount.
-    :type previous_awards:  List[models.Award]
-
-    :param lang: The lang requested.
-    :type lang: str
-
-    :return: The generated table.
-    :rtype: Table
-    """
-
-    payment_method_text = f"""Habilita Pago Adelantado: {
-    format_currency(award.payment_method.get("habilita_pago_adelantado", ""), award.award_currency)
-}
-Valor De Pago Adelantado: {
-    format_currency(award.payment_method.get("valor_de_pago_adelantado", ""), award.award_currency)
-}
-Valor Facturado: {
-    format_currency(award.payment_method.get("valor_facturado", ""), award.award_currency)
-}
-Valor Pendiente De Pago: {
-    format_currency(award.payment_method.get("valor_pendiente_de_pago", ""), award.award_currency)
-}
-Valor Pagado: {
-    format_currency(award.payment_method.get("valor_pagado", ""), award.award_currency)
-}
-"""
-    secop_link = '<link href="' + award.source_url + '">' + award.source_url + "</link>"
-
-    data = [
-        [
-            get_translated_string("Award Data", lang),
-            get_translated_string("Data", lang),
-        ],
-        [
-            get_translated_string("View data in SECOP II", lang),
-            Paragraph(secop_link, styleN),
-        ],
-        [
-            get_translated_string("Award Title", lang),
-            Paragraph(award.title, styleN),
-        ],
-        [
-            get_translated_string("Contracting Process ID", lang),
-            Paragraph(award.contracting_process_id, styleN),
-        ],
-        [
-            get_translated_string("Award Description", lang),
-            Paragraph(award.description, styleN),
-        ],
-        [
-            get_translated_string("Award Date", lang),
-            format_date(str(award.award_date)),
-        ],
-        [
-            get_translated_string("Award Value Currency & Amount", lang),
-            format_currency(award.award_amount, award.award_currency),
-        ],
-        [
-            get_translated_string("Contract Start Date", lang),
-            format_date(str(award.contractperiod_startdate)),
-        ],
-        [
-            get_translated_string("Contract End Date", lang),
-            format_date(str(award.contractperiod_enddate)),
-        ],
-        [
-            get_translated_string("Payment Method", lang),
-            payment_method_text,
-        ],
-        [
-            get_translated_string("Buyer Name", lang),
-            Paragraph(
-                award.buyer_name,
-                styleN,
-            ),
-        ],
-        [
-            get_translated_string("Procurement Method", lang),
-            Paragraph(award.procurement_method, styleN),
-        ],
-        [
-            get_translated_string("Contract Type", lang),
-            Paragraph(award.procurement_category, styleN),
-        ],
-    ]
-
-    return create_table(data)
